Secrets Manager - Additional tests (#4613)
* Secrets Manager - Additional tests

* Secrets Manager - Additional tests

+++ ibm/service/secretsmanager/resource_ibm_sm_private_certificate_configuration_intermediate_ca.go
@@ -806,7 +806,7 @@ func resourceIbmSmConfigurationActionPrivateCertificateSignIntermediateCAMapToCo
 		model.TTL = core.StringPtr(d.Get("ttl").(string))
 	}
 	if _, ok := d.GetOk("max_path_length"); ok {
-		model.MaxPathLength = core.Int64Ptr(d.Get("max_path_length").(int64))
+		model.MaxPathLength = core.Int64Ptr(int64(d.Get("max_path_length").(int)))
 	}
 	if _, ok := d.GetOk("exclude_cn_from_sans"); ok {
 		model.ExcludeCnFromSans = core.BoolPtr(d.Get("exclude_cn_from_sans").(bool))
